cafe-fausse-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
pydantic==2.8.2
Flask-Cors==4.0.1